pyproj/requirements.txt
Django==5.2.7
python-dotenv==1.1.1
dj-database-url==3.0.1
django-dbbackup==5.0.0
# requests==2.32.5
djlint==1.36.4
black==25.9.0
pytest-django==4.11.1
django-login-required-middleware==0.9.0
django-debug-toolbar==6.0.0
django-extensions==4.1
factory_boy==3.3.3
